UC WebDev_Final - Website (12/04/2022)
My final project(website) I worked on for the UCinci Fundamentals of Web Development Course. (As of 12/04/2022)

This is a website that is meant to showcase me and my skills as a candidate for an IT or Computer-Science related job or career.
This website is also meant to replace an old Google Site.